refactor(payment): reorganize payment transaction actions and update imports
