+++ apps/public/actions/upwithcrowd/payment-transaction/action.ts
@@ -1,14 +1,14 @@
 "use server";
-import {structuredError, structuredResponse} from "@repo/utils/api";
+import type {GetApiPaymentTransactionData} from "@ayasofyazilim/upwithcrowd-saas/UPWCService";
+import {structuredError, structuredSuccessResponse} from "@repo/utils/api";
 import {getUpwithcrowd} from "@/utils/client";
-import {GetApiPaymentTransactionData} from "@ayasofyazilim/upwithcrowd-saas/UPWCService";
 
 export async function getApiPaymentTransactionApi(data?: GetApiPaymentTransactionData) {
   try {
     const client = await getUpwithcrowd();
     const dataResponse = await client.paymentTransaction.getApiPaymentTransaction(data);
-    return structuredResponse(dataResponse);
+    return structuredSuccessResponse(dataResponse);
   } catch (error) {
-    return structuredError(error);
+    throw structuredError(error);
   }
 }
