+++ apps/public/actions/upwithcrowd/payment-transaction/post-action.ts
@@ -1,12 +1,12 @@
 "use server";
-import type {PostApiFileData} from "@ayasofyazilim/upwithcrowd-saas/UPWCService";
+import type {PostApiPaymentTransactionData} from "@ayasofyazilim/upwithcrowd-saas/UPWCService";
 import {structuredError, structuredResponse} from "@repo/utils/api";
 import {getUpwithcrowd} from "@/utils/client";
 
-export async function postApiPaymentTransaction(data: PostApiFileData) {
+export async function postApiPaymentTransaction(data: PostApiPaymentTransactionData) {
   try {
     const client = await getUpwithcrowd();
-    const dataResponse = await client.file.postApiFile(data);
+    const dataResponse = await client.paymentTransaction.postApiPaymentTransaction(data);
     return structuredResponse(dataResponse);
   } catch (error) {
     return structuredError(error);
